+++ CHANGELOG.md
@@ -8,7 +8,8 @@
  * Up Go requirement to v1.8+
  * s/failure/error/ in activity
  * Add `category` to experiment files
- * Change CLI to use a command structure. E.G. rulehunter serve --configdir=.
+ * Specify config filename rather than config directory through CLI
+ * Change CLI to use a command structure. E.G. rulehunter serve --config=.
  * Add version command to CLI
 
 ## 0.1 (7th July 2017)